Fix sensitivity type cast and add toggled value to sample app for validation

+++ camera_greenscreen/camera/android/src/main/java/io/flutter/plugins/camera/aardman/FilterParameters.java
@@ -27,9 +27,9 @@ public final class FilterParameters {
          if(colourDoubles != null){
              setReplacementColour(colourDoubles);
          }
-         Float thresholdSensitivity = (Float) arguments.get("sensitivity");
+         Double thresholdSensitivity = (Double) arguments.get("sensitivity");
          if (thresholdSensitivity != null){
-             sensitivity = thresholdSensitivity.floatValue();
+             sensitivity =  thresholdSensitivity.floatValue();
          }
          String backgroundImagePath = (String) arguments.get("backgroundPath");
          if(backgroundImagePath != null){
